+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 
 - Fix atomic floating-point min/max operations not returning the old value
   ([GH-1058](https://github.com/NVIDIA/warp/issues/1058)).
+- Fix a bug in CPU BVH construction when using groups for heterogenous edge cases where some groups have <= leaf size primitives ([GH-1111](https://github.com/NVIDIA/warp/issues/1111)).
 
 ## [1.10.1] - 2025-12-01